+++ resources/views/examples/audio-player/compact.blade.php
@@ -1,0 +1,1 @@
+<x-ui.audio-player src="https://www.soundhelix.com/examples/mp3/SoundHelix-Song-2.mp3" />