+++ resources/views/examples/audio-player/default.blade.php
@@ -1,0 +1,5 @@
+<x-ui.audio-player
+    src="https://www.soundhelix.com/examples/mp3/SoundHelix-Song-1.mp3"
+    title="SoundHelix Song 1"
+    artist="T. Schürger"
+/>