+++ resources/views/examples/image/default.blade.php
@@ -1,0 +1,5 @@
+<x-ui.image
+    src="https://picsum.photos/id/1015/600/400"
+    alt="A river winding through a rocky mountain valley"
+    class="w-[300px]"
+/>