+++ resources/views/examples/image/fallback.blade.php
@@ -1,0 +1,6 @@
+<x-ui.image
+    src="https://picsum.photos/this-image-does-not-exist.jpg"
+    alt="Quarterly revenue chart"
+    ratio="16/9"
+    class="w-[400px]"
+/>