+++ resources/views/examples/image/placeholder.blade.php
@@ -1,0 +1,7 @@
+<x-ui.image
+    src="https://picsum.photos/id/1018/800/450"
+    placeholder="https://picsum.photos/id/1018/20/11"
+    alt="A mountain lake reflecting a clear blue sky"
+    ratio="16/9"
+    class="w-[400px]"
+/>