+++ resources/views/examples/image/with-ratio.blade.php
@@ -1,0 +1,6 @@
+<x-ui.image
+    src="https://picsum.photos/id/1016/800/450"
+    alt="A misty forest seen from a mountain ridge"
+    ratio="16/9"
+    class="w-[400px]"
+/>