+++ resources/views/examples/map/default.blade.php
@@ -1,0 +1,6 @@
+<x-ui.map
+    class="max-w-2xl"
+    :lat="50.8467"
+    :lon="4.3499"
+    label="Brussels, Belgium"
+/>