+++ resources/views/examples/map/no-marker.blade.php
@@ -1,0 +1,8 @@
+<x-ui.map
+    class="max-w-2xl"
+    :lat="51.5079"
+    :lon="-0.0877"
+    :zoom="13"
+    :marker="false"
+    label="City of London"
+/>