+++ resources/views/examples/map/tall.blade.php
@@ -1,0 +1,8 @@
+<x-ui.map
+    class="max-w-2xl"
+    :lat="48.8584"
+    :lon="2.2945"
+    :zoom="16"
+    :height="480"
+    label="Eiffel Tower, Paris"
+/>